внесены правки после ревью
