+++ FakeNFT/Scenes /Catalog/CatalogScreen/Model/CatalogModel.swift
@@ -29,11 +29,3 @@ final class CatalogModel {
         }
     }
 }
-
-struct CatalogRequest: NetworkRequest {
-    
-    var endpoint: URL? {
-        URL(string: "\(RequestConstants.baseURL)/api/v1/collections")
-    }
-    var dto: Dto?
-}
