+++ FakeNFT/Scenes /Catalog/CatalogScreen/Model/NFTModelCatalog.swift
@@ -10,7 +10,7 @@ import UIKit
 struct NFTModelCatalog: Codable {
     let createdAt: String
     let name: String
-    let cover: String
+    let cover: URL
     let nfts: [String]
     let description: String
     let author: String
