+++ FakeNFT/Scenes /Catalog/CatalogScreen/View/CatalogViewController.swift
@@ -1,4 +1,5 @@
 import UIKit
+import ProgressHUD
 
 final class CatalogViewController: UIViewController {
     
@@ -61,8 +62,11 @@ final class CatalogViewController: UIViewController {
     }
     
     private func loadData() {
+        ProgressHUD.show()
+        ProgressHUD.animationType = .circleSpinFade
         viewModel.fetchCollections {
             DispatchQueue.main.async {
+                ProgressHUD.dismiss()
                 self.NFTTableView.reloadData()
             }
         }
@@ -115,6 +119,7 @@ extension CatalogViewController: UITableViewDataSource {
             return UITableViewCell()
         }
         let nft = viewModel.collection(at: indexPath.row)
+        cell.prepareForReuse()
         cell.configCell(name: nft.name, count: nft.nfts.count, image: nft.cover)
         cell.selectionStyle = .none
         return cell
